+++ changelog.md
@@ -44,7 +44,8 @@
 * support for function types added.
 * asm output for returning values and void functions made
 
-## 12/05/24 - directives, compiler options
+## 12/05/24 - directives, compiler options, externs
 
 * Added compiler directives for .well files, ``#NOMAIN`` for example tells the compiler to ignore main verification.
 * Added compiler CLI options for output names and object-only output.
+* Added extern support: ``~extern foo``.